requirements.txt
mako==1.3.5
pyyaml==6.0.1
black==24.8.0
